refactor: cleaning and puting directories at same level of abstraction
