--- backend-spring-boot/src/main/java/com/pharmacy/cpis/dto/JwtAuthenticationRequest.java
@@ -1,34 +1,0 @@
-package com.pharmacy.cpis.dto;
-
-// DTO for login
-public class JwtAuthenticationRequest {
-
-    private String email;
-    private String password;
-
-    public JwtAuthenticationRequest() {
-        super();
-    }
-
-    public JwtAuthenticationRequest(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-}
