--- backend-spring-boot/src/main/java/com/pharmacy/cpis/repository/IAuthorityRepository.java
@@ -1,9 +1,0 @@
-package com.pharmacy.cpis.repository;
-
-import com.pharmacy.cpis.model.Authority;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface IAuthorityRepository extends JpaRepository<Authority, Long> {
-    Authority findByName(String name);
-}
-
