--- backend-spring-boot/src/main/java/com/pharmacy/cpis/security/auth/RestAuthenticationEntryPoint.java
@@ -1,22 +1,0 @@
-package com.pharmacy.cpis.security.auth;
-
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class RestAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    // Method to be performed if (for forwarded credentials) the user tries to access a protected REST service
-    // Method returns 401 Unauthorized response, if there is a Login Page in the application, it is desirable that the user is redirected to that page
-    @Override
-    public void commence(HttpServletRequest request,
-                         HttpServletResponse response,
-                         AuthenticationException authException) throws IOException {
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, authException.getMessage());
-    }
-}
