--- backend-spring-boot/src/main/java/com/pharmacy/cpis/security/auth/TokenBasedAuthentication.java
@@ -1,40 +1,0 @@
-package com.pharmacy.cpis.security.auth;
-
-import org.springframework.security.authentication.AbstractAuthenticationToken;
-import org.springframework.security.core.userdetails.UserDetails;
-
-// Override authentication class instead of e.g. Basic Auth
-public class TokenBasedAuthentication extends AbstractAuthenticationToken {
-
-    private String token;
-    private final UserDetails principle;
-
-    public TokenBasedAuthentication(UserDetails principle) {
-        super(principle.getAuthorities());
-        this.principle = principle;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    @Override
-    public boolean isAuthenticated() {
-        return true;
-    }
-
-    @Override
-    public Object getCredentials() {
-        return token;
-    }
-
-    @Override
-    public UserDetails getPrincipal() {
-        return principle;
-    }
-
-}
