--- backend-spring-boot/src/main/java/com/pharmacy/cpis/service/IAuthorityService.java
@@ -1,12 +1,0 @@
-package com.pharmacy.cpis.service;
-
-import com.pharmacy.cpis.model.Authority;
-
-import java.util.List;
-
-public interface IAuthorityService {
-
-    List<Authority> findById(Long id);
-
-    List<Authority> findByName(String name);
-}
