--- backend-spring-boot/src/main/java/com/pharmacy/cpis/service/impl/AuthorityService.java
@@ -1,34 +1,0 @@
-package com.pharmacy.cpis.service.impl;
-
-import com.pharmacy.cpis.model.Authority;
-import com.pharmacy.cpis.repository.IAuthorityRepository;
-import com.pharmacy.cpis.service.IAuthorityService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Service
-public class AuthorityService implements IAuthorityService{
-
-    @Autowired
-    private IAuthorityRepository authorityRepository;
-
-    @Override
-    public List<Authority> findById(Long id) {
-        Authority auth = this.authorityRepository.getOne(id);
-        List<Authority> auths = new ArrayList<>();
-        auths.add(auth);
-        return auths;
-    }
-
-    @Override
-    public List<Authority> findByName(String name) {
-        Authority auth = this.authorityRepository.findByName(name);
-        List<Authority> auths = new ArrayList<>();
-        auths.add(auth);
-        return auths;
-    }
-
-}
