--- backend-spring-boot/src/main/java/com/pharmacy/cpis/service/impl/UserService.java
@@ -1,69 +1,0 @@
-package com.pharmacy.cpis.service.impl;
-
-import com.pharmacy.cpis.repository.IUserRepository;
-import com.pharmacy.cpis.service.IAuthorityService;
-import com.pharmacy.cpis.service.IUserService;
-import com.pharmacy.cpis.userservice.model.users.UserAccount;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class UserService implements IUserService, UserDetailsService {
-    @Autowired
-    private IUserRepository userRepository;
-
-    @Autowired
-    private IAuthorityService authService;
-
-    @Override
-    public UserAccount findOne(Long id) {
-        return userRepository.findById(id).orElseGet(null);
-    }
-
-    @Override
-    public List<UserAccount> findAll() {
-        return userRepository.findAll();
-    }
-
-    @Override
-    public UserAccount save(UserAccount userRequest) {
-//        UserAccount newUserAccount = new UserAccount();
-//
-//        newUserAccount.setEmail(userRequest.getEmail());
-//        newUserAccount.setPassword(userRequest.getPassword());  //TODO: Implement ENCODER/DECODER for password
-//        newUserAccount.setActive(false);                        //TODO: Change this to true when user activate account
-//        newUserAccount.setPerson(userRequest.getPerson());
-//
-//        List<Authority> auth = authService.findByName("ROLE_PATIENT");
-//        newUserAccount.setAuthorities(auth);
-//
-//        newUserAccount = userRepository.save(newUserAccount);
-        return userRepository.save(userRequest);
-    }
-
-    @Override
-    public void remove(Long id) {
-        userRepository.deleteById(id);
-    }
-
-    @Override
-    public UserAccount findByEmail(String email) {
-        return userRepository.findByEmail(email);
-    }
-
-    // Username is unique identifier in UserDetailsService, in our system that is email !
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        UserAccount user = userRepository.findByEmail(email);
-        if (user == null) {
-            throw new UsernameNotFoundException(String.format("No user found with email '%s'.", email));
-        } else {
-            return user;
-        }
-    }
-}
