+++ backend-spring-boot/src/main/java/com/pharmacy/cpis/userservice/dto/UserActivationDTO.java
@@ -1,0 +1,41 @@
+package com.pharmacy.cpis.userservice.dto;
+
+public class UserActivationDTO {
+    private String id;
+    private String email;
+    private String name;
+
+    public UserActivationDTO(){
+
+    }
+
+    public UserActivationDTO(String id, String email, String name) {
+        this.id = id;
+        this.email = email;
+        this.name = name;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
