+++ backend-spring-boot/src/main/java/com/pharmacy/cpis/userservice/dto/UserTokenState.java
@@ -1,0 +1,34 @@
+package com.pharmacy.cpis.userservice.dto;
+
+// A DTO that encapsulates the generated JWT and its duration that is returned to the client
+public class UserTokenState {
+
+    private String accessToken;
+    private Long expiresIn;
+
+    public UserTokenState() {
+        this.accessToken = null;
+        this.expiresIn = null;
+    }
+
+    public UserTokenState(String accessToken, long expiresIn) {
+        this.accessToken = accessToken;
+        this.expiresIn = expiresIn;
+    }
+
+    public String getAccessToken() {
+        return accessToken;
+    }
+
+    public void setAccessToken(String accessToken) {
+        this.accessToken = accessToken;
+    }
+
+    public Long getExpiresIn() {
+        return expiresIn;
+    }
+
+    public void setExpiresIn(Long expiresIn) {
+        this.expiresIn = expiresIn;
+    }
+}
