+++ backend-spring-boot/src/main/java/com/pharmacy/cpis/userservice/model/users/Authority.java
@@ -1,0 +1,46 @@
+package com.pharmacy.cpis.userservice.model.users;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import org.springframework.security.core.GrantedAuthority;
+
+import javax.persistence.*;
+
+// POJO which implement Spring Security GrantedAuthority, with that we can define roles in application
+@Entity
+@Table(name="AUTHORITY")
+public class Authority implements GrantedAuthority {
+
+    private static final long serialVersionUID = 1L;
+
+    @Id
+    @Column(name="id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    Long id;
+
+    @Column(name="name")
+    String name;
+
+    @Override
+    public String getAuthority() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @JsonIgnore
+    public String getName() {
+        return name;
+    }
+
+    @JsonIgnore
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+}
