+++ backend-spring-boot/src/main/java/com/pharmacy/cpis/userservice/repository/IPersonRepository.java
@@ -1,0 +1,7 @@
+package com.pharmacy.cpis.userservice.repository;
+
+import com.pharmacy.cpis.userservice.model.users.Person;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IPersonRepository extends JpaRepository<Person, Long> {
+}
