+++ backend-spring-boot/src/main/java/com/pharmacy/cpis/userservice/service/IAuthorityService.java
@@ -1,0 +1,12 @@
+package com.pharmacy.cpis.userservice.service;
+
+import com.pharmacy.cpis.userservice.model.users.Authority;
+
+import java.util.List;
+
+public interface IAuthorityService {
+
+    List<Authority> findById(Long id);
+
+    List<Authority> findByName(String name);
+}
