+++ backend-spring-boot/src/main/java/com/pharmacy/cpis/userservice/service/IUserService.java
@@ -1,0 +1,18 @@
+package com.pharmacy.cpis.userservice.service;
+
+import com.pharmacy.cpis.userservice.model.users.UserAccount;
+
+import java.util.List;
+
+public interface IUserService {
+
+    UserAccount findOne(Long id);
+
+    List<UserAccount> findAll();
+
+    UserAccount save(UserAccount userRequest);
+
+    void remove(Long id);
+
+    UserAccount findByEmail(String email);
+}
